// frontend/src/api/models/dtoFlagResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { DtoFlagValueDTO } from './dtoFlagValueDTO'
import type { DtoTargetingRuleDTO } from './dtoTargetingRuleDTO'

export interface DtoFlagResponse {
  created_at?: string
  created_by?: string
  default_value?: DtoFlagValueDTO
  description?: string
  id?: string
  key?: string
  name?: string
  rules?: DtoTargetingRuleDTO[]
  status?: string
  tags?: string[]
  type?: string
  updated_at?: string
  updated_by?: string
  version?: number
}