// frontend/src/api/models/listFeatureFlagFlagsParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { ListFeatureFlagFlagsStatus } from './listFeatureFlagFlagsStatus'
import type { ListFeatureFlagFlagsType } from './listFeatureFlagFlagsType'

export type ListFeatureFlagFlagsParams = {
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
  /**
   * Filter by status
   */
  status?: ListFeatureFlagFlagsStatus
  /**
   * Filter by type
   */
  type?: ListFeatureFlagFlagsType
  /**
   * Filter by tags (comma-separated)
   */
  tags?: string
  /**
   * Search term
   */
  search?: string
}