// frontend/src/api/models/handlerCreateFlagHTTPRequestType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerCreateFlagHTTPRequestType =
  (typeof HandlerCreateFlagHTTPRequestType)[keyof typeof HandlerCreateFlagHTTPRequestType]

export const HandlerCreateFlagHTTPRequestType = {
  boolean: 'boolean',
  percentage: 'percentage',
  variant: 'variant',
  user_segment: 'user_segment',
} as const