// frontend/src/api/models/listFeatureFlagFlagsStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListFeatureFlagFlagsStatus =
  (typeof ListFeatureFlagFlagsStatus)[keyof typeof ListFeatureFlagFlagsStatus]

export const ListFeatureFlagFlagsStatus = {
  enabled: 'enabled',
  disabled: 'disabled',
  archived: 'archived',
} as const